pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "code-review-env"
version = "1.0.0"
description = "OpenEnv-compatible AI code review training environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
]
inference = [
    "openai>=1.0",
]
dev = [
    "pytest>=8.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "openai>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["code_review_env*", "server*"]